+++ blueprints/__pycache__/curvatures.cpython-311.pyc
