+++ blueprints/__pycache__/fundamental_form.cpython-311.pyc
