+++ blueprints/__pycache__/representacion.cpython-311.pyc
