+++ blueprints/caracteristicas.py
@@ -1,31 +1,7 @@
 from flask import Blueprint, request
-import sympy as sp
+from .utils import normaliza_parametrizacion
 
 caracteristicas_bp = Blueprint('caracteristicas', __name__)
-
-def normaliza_parametrizacion(var1, var2, sup):
-    if var1:
-        u = sp.symbols(var1, real = True)
-    else:
-        u = sp.symbols('u', real = True)
-        var1='u'
-
-    if var2:
-        v = sp.symbols(var2, real = True)
-    else:
-        v = sp.symbols('v', real = True)
-        var2='v'
-
-    elementos_str = sup.strip('[]').split(',')
-    try:
-        superficie = [sp.sympify(elem, locals={var1: u, var2: v}) for elem in elementos_str]
-    except Exception as e:
-        raise Exception(f"Error al procesar la superficie: {e}")
-    if len(superficie) != 3:
-        raise Exception(f"La parametrización de la superficie debe tener 3 elementos pero se han encontrado {len(superficie)}")
-    
-    return superficie, u, v
-
 
 @caracteristicas_bp.route('/vector_normal')
 def vector_normal():
