+++ blueprints/representacion.py
